pyproject.toml
[project]
name = "agent-learn"
version = "0.1.0"
description = "LangChain & LangGraph 系统学习项目"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "faiss-cpu>=1.13.2",
    "langchain>=1.2.13",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.22",
    "langchain-openai>=1.1.12",
    "langgraph>=1.1.3",
    "python-dotenv>=1.0.0",
]
